cherry-pick upstream:
[process] Refactor user stack/args functions into portable code

...

[process][riscv] Share arch_specific task setup/free funcs with RISCV

[riscv] Access regs_t members using portable inline function
