+++ .vscode/settings.json
@@ -221,6 +221,9 @@
       "compare": "c",
       "cmdline_opts.h": "c",
       "mod_kb8042.h": "c",
-      "process_mm.h": "c"
+      "process_mm.h": "c",
+      "hal_types.h": "c",
+      "vdso.h": "c",
+      "image.h": "c"
    },
 }
