+++ include/tilck/kernel/arch/x86_64/arch_utils.h
@@ -14,6 +14,8 @@ struct x86_64_regs {
    s32 int_num;
    u32 err_code;
    u64 rsp, rbp, rax, rip;
+   u64 rflags;
+   u64 userrsp;
    u32 custom_flags;
 };
 
@@ -24,7 +26,7 @@ struct x86_64_arch_proc_members {
 
 struct x86_64_arch_task_members {
    u16 fpu_regs_size;
-   void *aligned_fpu_regs;
+   void *fpu_regs;
 };
 
 static ALWAYS_INLINE int regs_intnum(regs_t *r)
@@ -35,7 +37,12 @@ static ALWAYS_INLINE int regs_intnum(regs_t *r)
 
 static ALWAYS_INLINE void set_return_register(regs_t *r, ulong value)
 {
-   NOT_IMPLEMENTED();
+   r->rax = value;
+}
+
+static ALWAYS_INLINE ulong get_return_register(regs_t *r)
+{
+   return r->rax;
 }
 
 NORETURN static ALWAYS_INLINE void context_switch(regs_t *r)
@@ -61,4 +68,19 @@ static ALWAYS_INLINE void *regs_get_frame_ptr(regs_t *r)
 static ALWAYS_INLINE void *regs_get_ip(regs_t *r)
 {
    return TO_PTR(r->rip);
+}
+
+static ALWAYS_INLINE void regs_set_ip(regs_t *r, ulong value)
+{
+   r->rip = value;
+}
+
+static ALWAYS_INLINE ulong regs_get_usersp(regs_t *r)
+{
+   return r->userrsp;
+}
+
+static ALWAYS_INLINE void regs_set_usersp(regs_t *r, ulong value)
+{
+   r->userrsp = value;
 }
