+++ include/tilck/kernel/hal.h
@@ -7,6 +7,8 @@
 #include <tilck/common/datetime.h>
 #include <tilck/kernel/hal_types.h>
 
+#define USERMODE_STACK_ALIGN              16ul
+
 #if defined(__i386__) || defined(__x86_64__)
 
    #define arch_x86_family
