+++ kernel/arch/x86_64/process64.c
@@ -24,14 +24,8 @@
 
 #include <tilck/mods/tracing.h>
 
-bool
-arch_specific_new_task_setup(struct task *ti, struct task *parent)
-{
-   NOT_IMPLEMENTED();
-}
-
 void
-arch_specific_free_task(struct task *ti)
+setup_usermode_task_regs(regs_t *r, void *entry, void *stack_addr)
 {
    NOT_IMPLEMENTED();
 }
@@ -60,42 +54,11 @@ switch_to_task(struct task *ti)
    NOT_IMPLEMENTED();
 }
 
-void
-finalize_usermode_task_setup(struct task *ti, regs_t *user_regs)
-{
-   NOT_IMPLEMENTED();
-}
-
-int setup_process(struct elf_program_info *pinfo,
-                  struct task *ti,
-                  const char *const *argv,
-                  const char *const *env,
-                  struct task **ti_ref,
-                  regs_t *r)
-{
-   NOT_IMPLEMENTED();
-}
-
-void setup_pause_trampoline(regs_t *r)
-{
-   NOT_IMPLEMENTED();
-}
-
 int setup_sig_handler(struct task *ti,
                       enum sig_state sig_state,
                       regs_t *r,
                       ulong user_func,
                       int signum)
-{
-   NOT_IMPLEMENTED();
-}
-
-void task_info_reset_kernel_stack(struct task *ti)
-{
-   NOT_IMPLEMENTED();
-}
-
-void save_current_task_state(regs_t *r, bool irq)
 {
    NOT_IMPLEMENTED();
 }
